feeder v0.0.1 Feeder.Discovery

Feeder.Discovery aims xml parser library to find proper url for API and also proper mapping to match our db model.

Functions

map(service, options \\ :empty)

Associate a xml mapping with the service asked

returns a map

## Example

iex> Feeder.Discovery.map({"test", "odds"})
  [matchs: [%SweetXpath{cast_to: false, is_keyword: false, is_list: true,
     is_optional: false, is_value: true, path: '//Match'},
    {:match_id,
     %SweetXpath{cast_to: false, is_keyword: false, is_list: false,
      is_optional: false, is_value: true, path: './@Id'}}]]

url(service, options \\ nil)

Associate a url with the service asked

return a binary

## Example

iex> Feeder.Discovery.url({"test", "odds"})
  "https://hello_ca_marche"

  iex> Feeder.Discovery.url({"winamax", "odds"}, %{sport_number: 1})
  "https://up.winamax.fr/partners/betting/betteur/export.php?sport=1"