feeder v0.0.1 Feeder.Parser

This module aims to fetch, parse and insert or update xml third party API resource into db

Functions

fetch(services, options \\ nil)

In charge of making the RPC to third party API then passing the response to our handler and to our xml Parser

== Example

iex> Feeder.Parser.fetch({"test", "odds"})
{:error, "nil"}

handle_response(arg)

handle the response from HTTPoison call to remote resource

## Example

iex> Feeder.Parser.handle_response({:ok, %{status_code: 200, body: "<xml></xml>"}})
 {:ok, "<xml></xml>"}

 iex> Feeder.Parser.handle_response({:error, %{status_code: 500, body: "<xml></xml>"}})
 {:error, "<xml></xml>"}

 iex> Feeder.Parser.handle_response({:error, %HTTPoison.Error{id: 30, reason: "not nice"}})
 {:error, "nil"}

parse_xml(arg, service, options \\ nil)

Parse Xml given a certain body, services and options

Example

iex> Feeder.Parser.parse_xml({:ok, "<xml><Match Id='21312'></Match></xml>"}, {"test", "odds"})
{:ok, %{matchs: [%{match_id: '21312'}]}}

iex> Feeder.Parser.parse_xml({:error, "fail"}, {"test", "odds"})
{:error, "fail"}